simpleLogFilter tool to perform log filtering
This tool can be used by any user who wants to filter a big text or log file without having to code. The tool creates an output file as per user's request with the filtered results.
The user has the option to enter keywords or tags. The tool also provides suggestions for the keywords using different separators. The tool is expected to work in both Windows and Linux environment.