Updated with the Readme for simpleUMLGenerator
The tool simpleUMLGenerator utilizes the pyreverse module of pylint along with the GraphViz software. This tool provides a simple UI for the user to select the Python module, output location and desired output format. All the work is done by the pyreverse module and dot format present in the GraphViz.